add/delete course

+++ server/src/main/java/org/coders/ratemycourses/service/CourseService.java
@@ -1,6 +1,7 @@
 package org.coders.ratemycourses.service;
 
 import java.util.List;
+import java.util.Optional;
 
 import org.coders.ratemycourses.model.Course;
 import org.coders.ratemycourses.model.Department;
@@ -14,12 +15,20 @@ public class CourseService{
     CourseRepo repo;
 
     public void addNewCourse(Course theCourse){
-        repo.save(theCourse);
+        try {
+            if( !theCourse.getDepartmentId().isEmpty()){
+                repo.save(theCourse);
+            }
+        } catch (Exception e) {
+            e.printStackTrace();
+        }  
     }
 
     public void deleteCourse(String id){
         try {
+            if( !repo.findById(id).isEmpty()){
             repo.deleteById(id);
+            }
 
         } catch (Exception e) {
             e.printStackTrace();
